pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bci-snn-rl"
version = "0.1.0"
description = "Research infra: MI EEG -> closed-loop / continuous control with SNN + RL"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
  "pyyaml>=6.0",
  "numpy>=1.24",
  "torch>=2.1",
  "scikit-learn>=1.3",
  "gymnasium>=0.29",
  "matplotlib>=3.8",
  "mne>=1.6",
  "moabb>=1.2.0",
  "spikingjelly>=0.0.0.0.14",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

